+++ backend/Controllers/RecruitmentsController.cs
@@ -1,16 +1,14 @@
-using System.Data;
 using ClubHub.Api.Data;
 using ClubHub.Api.Data.Entities;
+using ClubHub.Api.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Storage;
 using CreateRecruitmentApplicationRequest = Org.OpenAPITools.Models.CreateRecruitmentApplicationRequest;
 using CreateRecruitmentRequest = Org.OpenAPITools.Models.CreateRecruitmentRequest;
-using RecruitmentApplicationDto = Org.OpenAPITools.Models.RecruitmentApplication;
-using RecruitmentDto = Org.OpenAPITools.Models.Recruitment;
 using ReviewRecruitmentApplicationRequest = Org.OpenAPITools.Models.ReviewRecruitmentApplicationRequest;
 using ReviewRecruitmentRequest = Org.OpenAPITools.Models.ReviewRecruitmentRequest;
 using UpdateRecruitmentRequest = Org.OpenAPITools.Models.UpdateRecruitmentRequest;
+using static ClubHub.Api.Services.RecruitmentWorkflow;
 
 namespace ClubHub.Api.Controllers;
 
@@ -18,34 +16,14 @@ namespace ClubHub.Api.Controllers;
 [Route("api/[controller]")]
 public class RecruitmentsController : ControllerBase
 {
-    private const string ApplicationPending = "pending";
-    private const string ApplicationAccepted = "accepted";
-    private const string ApplicationRejected = "rejected";
-    private const string ReviewApproved = "approved";
-    private const string ReviewRejected = "rejected";
-    private const string ClubApproved = "approved";
-    private const string ClubActive = "active";
-    private const string MemberActive = "active";
-    private const string ClubMemberRoleCode = "CLUB_MEMBER";
-    private const int MaxStudentClubMemberships = 3;
-
-    private static readonly TimeZoneInfo BusinessTimeZone = ResolveBusinessTimeZone();
-    private static readonly HashSet<string> RecruitmentManagerRoleCodes = new(StringComparer.OrdinalIgnoreCase)
-    {
-        "club_officer",
-        "club_leader"
-    };
-    private static readonly HashSet<string> ClubMembershipRoleCodes = new(StringComparer.OrdinalIgnoreCase)
-    {
-        ClubMemberRoleCode,
-        "CLUB_OFFICER",
-        "CLUB_LEADER",
-        "club_president"
-    };
-
     private readonly ClubHubDbContext _db;
+    private readonly RecruitmentApplicationService _applicationService;
 
-    public RecruitmentsController(ClubHubDbContext db) => _db = db;
+    public RecruitmentsController(ClubHubDbContext db, RecruitmentApplicationService applicationService)
+    {
+        _db = db;
+        _applicationService = applicationService;
+    }
 
     [HttpGet]
     public async Task<IActionResult> GetAll(
@@ -285,212 +263,27 @@ public class RecruitmentsController : ControllerBase
     [HttpGet("{recruitId:int}/applications")]
     public async Task<IActionResult> GetApplications(int recruitId, [FromQuery] int viewerUserId)
     {
-        if (viewerUserId <= 0) return BadRequest(new { message = "请选择当前用户。" });
-
-        var viewer = await LoadUserAsync(viewerUserId);
-        if (viewer is null) return NotFound(new { message = "当前用户不存在。" });
-
-        var recruitment = await RecruitmentQuery(asNoTracking: true).FirstOrDefaultAsync(r => r.RecruitId == recruitId);
-        if (recruitment is null) return NotFound(new { message = "招募不存在。" });
-
-        var canManage = CanManageRecruitment(viewer, recruitment.ClubId);
-        var query = ApplicationQuery(asNoTracking: true).Where(a => a.RecruitId == recruitId);
-        if (!canManage)
-        {
-            query = query.Where(a => a.UserId == viewer.UserId);
-        }
-
-        var applications = await query
-            .OrderByDescending(a => a.SubmittedAt)
-            .ThenByDescending(a => a.ApplicationId)
-            .ToListAsync();
-
-        if (!canManage && applications.Count == 0)
-        {
-            return StatusCode(403, new { message = "当前用户没有查看该招募报名的权限。" });
-        }
-
-        return Ok(applications.Select(ToApplicationDto));
+        var result = await _applicationService.GetApplicationsAsync(recruitId, viewerUserId);
+        return ToActionResult(result);
     }
 
     [HttpPost("{recruitId:int}/applications")]
     public async Task<IActionResult> CreateApplication(int recruitId, [FromBody] CreateRecruitmentApplicationRequest req)
     {
-        if (req.CurrentUserId <= 0) return BadRequest(new { message = "请选择当前报名用户。" });
-        if (string.IsNullOrWhiteSpace(req.ApplicationReason)) return BadRequest(new { message = "报名理由不能为空。" });
+        var result = await _applicationService.CreateApplicationAsync(recruitId, req);
+        if (!result.Succeeded) return ToActionResult(result);
 
-        var applicant = await LoadUserAsync(req.CurrentUserId);
-        if (applicant is null) return NotFound(new { message = "当前用户不存在。" });
-        if (!UsersController.IsActive(applicant.AccountStatus))
-        {
-            return BadRequest(new { message = "当前用户账号不可用，不能提交招募报名。" });
-        }
-        if (!UsersController.IsStudent(applicant) || UsersController.IsPlatformAdmin(applicant))
-        {
-            return StatusCode(403, new { message = "只有普通学生可以提交招募报名。" });
-        }
-
-        await using var transaction = await _db.Database.BeginTransactionAsync(IsolationLevel.Serializable);
-        if (!await LockRecruitmentRowAsync(recruitId))
-        {
-            return NotFound(new { message = "招募不存在。" });
-        }
-
-        var recruitment = await RecruitmentQuery().FirstOrDefaultAsync(r => r.RecruitId == recruitId);
-        if (recruitment is null) return NotFound(new { message = "招募不存在。" });
-        if (recruitment.Club is null || !IsMaintainableClub(recruitment.Club))
-        {
-            return Conflict(new { message = "社团状态不允许接收报名。" });
-        }
-        var now = BusinessNow();
-        if (EffectiveRecruitmentStatus(recruitment, now) != RecruitmentStatuses.Accepting)
-        {
-            return Conflict(new { message = "只有申请中的纳新可以提交报名。" });
-        }
-
-        var hasSubmitted = await _db.RecruitmentApplications.AnyAsync(a =>
-            a.RecruitId == recruitId &&
-            a.UserId == applicant.UserId);
-        if (hasSubmitted) return Conflict(new { message = "你已经提交过该招募报名，请勿重复提交。" });
-
-        if (await IsCurrentClubMemberAsync(recruitment.ClubId, applicant.UserId))
-        {
-            return Conflict(new { message = "你已经是该社团成员，无需再次报名招募。" });
-        }
-
-        var currentClubCount = await CountCurrentMembershipClubsAsync(applicant.UserId);
-        if (currentClubCount >= MaxStudentClubMemberships)
-        {
-            return Conflict(new { message = "一个学生最多只能同时加入 3 个社团，当前已达到上限。" });
-        }
-
-        if (recruitment.Quota is not null && recruitment.Applications.Count(a => a.ApplicationStatus == ApplicationAccepted) >= recruitment.Quota.Value)
-        {
-            return Conflict(new { message = "招募名额已满，暂时不能继续报名。" });
-        }
-
-        var nextId = (await _db.RecruitmentApplications.MaxAsync(a => (int?)a.ApplicationId) ?? 0) + 1;
-        var application = new RecruitmentApplication
-        {
-            ApplicationId = nextId,
-            RecruitId = recruitment.RecruitId,
-            UserId = applicant.UserId,
-            ApplicationReason = req.ApplicationReason.Trim(),
-            ApplicationStatus = ApplicationPending,
-            SubmittedAt = DateTime.UtcNow,
-            Recruitment = recruitment,
-            User = applicant
-        };
-
-        _db.RecruitmentApplications.Add(application);
-        await _db.SaveChangesAsync();
-        await transaction.CommitAsync();
-
-        var created = await ApplicationQuery().FirstAsync(a => a.ApplicationId == application.ApplicationId);
-        return CreatedAtAction(nameof(GetApplications), new { recruitId, viewerUserId = req.CurrentUserId }, ToApplicationDto(created));
+        return CreatedAtAction(
+            nameof(GetApplications),
+            new { recruitId, viewerUserId = req.CurrentUserId },
+            result.Value);
     }
 
     [HttpPatch("applications/{applicationId:int}/review")]
     public async Task<IActionResult> ReviewApplication(int applicationId, [FromBody] ReviewRecruitmentApplicationRequest req)
     {
-        if (req.CurrentUserId <= 0) return BadRequest(new { message = "请选择当前筛选用户。" });
-
-        var decision = NormalizeApplicationStatus(req.Decision);
-        if (decision is not ApplicationAccepted and not ApplicationRejected)
-        {
-            return BadRequest(new { message = "筛选结果只能是 accepted 或 rejected。" });
-        }
-
-        if (req.InterviewScore is not null && (req.InterviewScore < 0 || req.InterviewScore > 100))
-        {
-            return BadRequest(new { message = "面试分数必须在 0 到 100 之间。" });
-        }
-
-        var reviewer = await LoadUserAsync(req.CurrentUserId);
-        if (reviewer is null) return NotFound(new { message = "当前用户不存在。" });
-
-        var application = await ApplicationQuery().FirstOrDefaultAsync(a => a.ApplicationId == applicationId);
-        if (application is null) return NotFound(new { message = "招募报名不存在。" });
-        if (application.Recruitment is null) return NotFound(new { message = "招募不存在。" });
-
-        if (!CanManageRecruitment(reviewer, application.Recruitment.ClubId))
-        {
-            return StatusCode(403, new { message = "只有系统管理员或本社团干部可以筛选招募报名。" });
-        }
-
-        if (application.ApplicationStatus != ApplicationPending)
-        {
-            return Conflict(new { message = "只有待筛选的报名可以录入筛选结果。" });
-        }
-
-        if (decision == ApplicationAccepted)
-        {
-            await using var transaction = await _db.Database.BeginTransactionAsync(IsolationLevel.Serializable);
-            if (!await LockRecruitmentRowAsync(application.RecruitId))
-            {
-                return NotFound(new { message = "招募不存在。" });
-            }
-
-            await _db.Entry(application).ReloadAsync();
-            if (application.Recruitment is null)
-            {
-                await _db.Entry(application).Reference(a => a.Recruitment).LoadAsync();
-            }
-
-            if (application.Recruitment is null)
-            {
-                return NotFound(new { message = "招募不存在。" });
-            }
-
-            if (application.ApplicationStatus != ApplicationPending)
-            {
-                return Conflict(new { message = "只有待筛选的报名可以录入筛选结果。" });
-            }
-
-            if (await IsCurrentClubMemberAsync(application.Recruitment.ClubId, application.UserId))
-            {
-                return Conflict(new { message = "该学生已经是社团成员，不能重复录取。" });
-            }
-
-            var currentClubCount = await CountCurrentMembershipClubsAsync(application.UserId);
-            if (currentClubCount >= MaxStudentClubMemberships)
-            {
-                return Conflict(new { message = "一个学生最多只能同时加入 3 个社团，该学生已达到上限。" });
-            }
-
-            var acceptedCount = await _db.RecruitmentApplications.CountAsync(a =>
-                a.RecruitId == application.RecruitId &&
-                a.ApplicationStatus == ApplicationAccepted);
-            if (application.Recruitment.Quota is not null && acceptedCount >= application.Recruitment.Quota.Value)
-            {
-                return Conflict(new { message = "招募名额已满，不能继续录取。" });
-            }
-
-            await AddAcceptedMemberAsync(application, DateTime.UtcNow);
-
-            application.ApplicationStatus = decision;
-            application.InterviewScore = req.InterviewScore is null ? null : Convert.ToDecimal(req.InterviewScore.Value);
-            application.ReviewerUserId = reviewer.UserId;
-            application.Reviewer = reviewer;
-            application.ReviewedAt = DateTime.UtcNow;
-
-            await _db.SaveChangesAsync();
-            await transaction.CommitAsync();
-
-            var accepted = await ApplicationQuery().FirstAsync(a => a.ApplicationId == application.ApplicationId);
-            return Ok(ToApplicationDto(accepted));
-        }
-
-        application.ApplicationStatus = decision;
-        application.InterviewScore = req.InterviewScore is null ? null : Convert.ToDecimal(req.InterviewScore.Value);
-        application.ReviewerUserId = reviewer.UserId;
-        application.Reviewer = reviewer;
-        application.ReviewedAt = DateTime.UtcNow;
-
-        await _db.SaveChangesAsync();
-
-        var reviewed = await ApplicationQuery().FirstAsync(a => a.ApplicationId == application.ApplicationId);
-        return Ok(ToApplicationDto(reviewed));
+        var result = await _applicationService.ReviewApplicationAsync(applicationId, req);
+        return ToActionResult(result);
     }
 
     private IQueryable<Recruitment> RecruitmentQuery(bool asNoTracking = false)
@@ -500,38 +293,6 @@ public class RecruitmentsController : ControllerBase
             .Include(r => r.Applications);
 
         return asNoTracking ? query.AsNoTracking() : query;
-    }
-
-    private IQueryable<RecruitmentApplication> ApplicationQuery(bool asNoTracking = false)
-    {
-        var query = _db.RecruitmentApplications
-            .Include(a => a.Recruitment)
-                .ThenInclude(r => r!.Club)
-            .Include(a => a.User)
-            .Include(a => a.Reviewer);
-
-        return asNoTracking ? query.AsNoTracking() : query;
-    }
-
-    private async Task<bool> LockRecruitmentRowAsync(int recruitId)
-    {
-        var connection = _db.Database.GetDbConnection();
-        if (connection.State != ConnectionState.Open)
-        {
-            await connection.OpenAsync();
-        }
-
-        await using var command = connection.CreateCommand();
-        command.Transaction = _db.Database.CurrentTransaction?.GetDbTransaction();
-        command.CommandText = "SELECT RECRUIT_ID FROM RECRUITMENTS WHERE RECRUIT_ID = :recruitId FOR UPDATE";
-
-        var parameter = command.CreateParameter();
-        parameter.ParameterName = "recruitId";
-        parameter.Value = recruitId;
-        command.Parameters.Add(parameter);
-
-        var lockedRecruitId = await command.ExecuteScalarAsync();
-        return lockedRecruitId is not null && lockedRecruitId != DBNull.Value;
     }
 
     private async Task<User?> LoadUserAsync(int userId) =>
@@ -555,464 +316,10 @@ public class RecruitmentsController : ControllerBase
             r.StartAt.Value < endAt &&
             r.EndAt.Value > startAt);
 
-    private async Task<bool> IsCurrentClubMemberAsync(int clubId, int userId)
+    private IActionResult ToActionResult<T>(ServiceResult<T> result)
     {
-        var today = BusinessToday();
-        var currentMemberRecordExists = await _db.ClubMembers.AnyAsync(m =>
-            m.ClubId == clubId &&
-            m.UserId == userId &&
-            (m.MemberStatus == null || m.MemberStatus == MemberActive) &&
-            (m.TermStart == null || m.TermStart <= today) &&
-            (m.TermEnd == null || m.TermEnd >= today));
-        if (currentMemberRecordExists) return true;
+        if (result.Succeeded) return Ok(result.Value);
 
-        var roleAssignments = await _db.UserRoles
-            .Include(ur => ur.Role)
-            .Where(ur => ur.UserId == userId && ur.ClubId == clubId)
-            .ToListAsync();
-        return roleAssignments.Any(ur => IsClubMembershipRole(ur.Role));
+        return StatusCode(result.StatusCode, new { message = result.ErrorMessage });
     }
-
-    private async Task<int> CountCurrentMembershipClubsAsync(int userId, int? excludingClubId = null)
-    {
-        var today = BusinessToday();
-        var memberClubIds = await _db.ClubMembers
-            .Where(m =>
-                m.UserId == userId &&
-                (excludingClubId == null || m.ClubId != excludingClubId.Value) &&
-                (m.MemberStatus == null || m.MemberStatus == MemberActive) &&
-                (m.TermStart == null || m.TermStart <= today) &&
-                (m.TermEnd == null || m.TermEnd >= today))
-            .Select(m => m.ClubId)
-            .ToListAsync();
-
-        var roleAssignments = await _db.UserRoles
-            .Include(ur => ur.Role)
-            .Where(ur =>
-                ur.UserId == userId &&
-                ur.ClubId != null &&
-                (excludingClubId == null || ur.ClubId != excludingClubId.Value))
-            .ToListAsync();
-
-        return memberClubIds
-            .Concat(roleAssignments
-                .Where(ur => IsClubMembershipRole(ur.Role))
-                .Select(ur => ur.ClubId!.Value))
-            .Distinct()
-            .Count();
-    }
-
-    private async Task AddAcceptedMemberAsync(RecruitmentApplication application, DateTime now)
-    {
-        if (application.Recruitment is null) return;
-
-        var nextMemberId = (await _db.ClubMembers.MaxAsync(m => (int?)m.MemberId) ?? 0) + 1;
-        _db.ClubMembers.Add(new ClubMember
-        {
-            MemberId = nextMemberId,
-            ClubId = application.Recruitment.ClubId,
-            UserId = application.UserId,
-            PositionName = "社员",
-            TermName = $"{BusinessDate(now).Year} 招新录取",
-            TermStart = BusinessDate(now),
-            MemberStatus = MemberActive,
-            JoinAt = now,
-            ContributionScore = 0
-        });
-
-        await EnsureClubMemberRoleAsync(application.Recruitment.ClubId, application.UserId, now);
-    }
-
-    private async Task EnsureClubMemberRoleAsync(int clubId, int userId, DateTime now)
-    {
-        var role = await EnsureClubRoleAsync(
-            ClubMemberRoleCode,
-            "社团成员",
-            "指定社团内角色，可查看社团内部信息、资源、通知和参与讨论、签到。",
-            now);
-
-        var hasRole = await _db.UserRoles.AnyAsync(ur =>
-            ur.UserId == userId &&
-            ur.ClubId == clubId &&
-            ur.Role != null &&
-            ur.Role.RoleCode.ToUpper() == ClubMemberRoleCode);
-        if (hasRole) return;
-
-        _db.UserRoles.Add(new UserRole
-        {
-            UserRoleId = await NextUserRoleIdAsync(),
-            UserId = userId,
-            RoleId = role.RoleId,
-            ClubId = clubId,
-            AssignedAt = now
-        });
-    }
-
-    private async Task<Role> EnsureClubRoleAsync(string roleCode, string roleName, string permissionDesc, DateTime now)
-    {
-        var role = await _db.Roles.FirstOrDefaultAsync(r => r.RoleCode.ToUpper() == roleCode);
-        if (role is not null) return role;
-
-        role = new Role
-        {
-            RoleId = await NextRoleIdAsync(),
-            RoleCode = roleCode,
-            RoleName = roleName,
-            RoleScope = "club",
-            PermissionDesc = permissionDesc,
-            CreatedAt = now
-        };
-        _db.Roles.Add(role);
-        return role;
-    }
-
-    private async Task<int> NextRoleIdAsync()
-    {
-        var maxSaved = await _db.Roles.MaxAsync(r => (int?)r.RoleId) ?? 0;
-        var maxAdded = _db.ChangeTracker.Entries<Role>()
-            .Where(entry => entry.State == EntityState.Added)
-            .Select(entry => entry.Entity.RoleId)
-            .DefaultIfEmpty(0)
-            .Max();
-
-        return Math.Max(maxSaved, maxAdded) + 1;
-    }
-
-    private async Task<int> NextUserRoleIdAsync()
-    {
-        var maxSaved = await _db.UserRoles.MaxAsync(ur => (int?)ur.UserRoleId) ?? 0;
-        var maxAdded = _db.ChangeTracker.Entries<UserRole>()
-            .Where(entry => entry.State == EntityState.Added)
-            .Select(entry => entry.Entity.UserRoleId)
-            .DefaultIfEmpty(0)
-            .Max();
-
-        return Math.Max(maxSaved, maxAdded) + 1;
-    }
-
-    private static RecruitmentDto ToRecruitmentDto(Recruitment recruitment, User viewer, DateTime now)
-    {
-        var currentApplication = recruitment.Applications
-            .Where(a => a.UserId == viewer.UserId)
-            .OrderByDescending(a => a.SubmittedAt)
-            .ThenByDescending(a => a.ApplicationId)
-            .FirstOrDefault();
-
-        var status = EffectiveRecruitmentStatus(recruitment, now);
-        var applicationStatus = NormalizeApplicationStatus(currentApplication?.ApplicationStatus);
-        var currentUserIsMember = CurrentUserIsMemberOfClub(viewer, recruitment.ClubId);
-        var isOwnProposal = IsOwnRecruitmentProposal(viewer, recruitment);
-
-        return new RecruitmentDto
-        {
-            Id = recruitment.RecruitId,
-            ClubId = recruitment.ClubId,
-            ClubName = recruitment.Club?.ClubName ?? $"社团 {recruitment.ClubId}",
-            Title = recruitment.Title,
-            Description = recruitment.Description,
-            StartAt = recruitment.StartAt,
-            EndAt = recruitment.EndAt,
-            Quota = recruitment.Quota,
-            Requirements = recruitment.Requirements,
-            RecruitStatus = ToRecruitmentStatusEnum(status),
-            RecruitStatusText = RecruitmentStatusText(status),
-            CreatedAt = recruitment.CreatedAt,
-            ApplicationCount = recruitment.Applications.Count,
-            AcceptedCount = recruitment.Applications.Count(a => a.ApplicationStatus == ApplicationAccepted),
-            CurrentUserApplicationId = currentApplication?.ApplicationId,
-            CurrentUserApplicationStatus = applicationStatus,
-            CurrentUserApplicationStatusText = applicationStatus is null ? null : ApplicationStatusText(applicationStatus),
-            CurrentUserIsMember = currentUserIsMember,
-            IsOwnProposal = isOwnProposal,
-            CanManage = CanManageRecruitment(viewer, recruitment.ClubId),
-            CanEdit = CanEditRecruitment(viewer, recruitment),
-            CanDelete = CanDeleteDraftRecruitment(viewer, recruitment),
-            CanReview = CanReviewRecruitment(viewer, recruitment)
-        };
-    }
-
-    private static RecruitmentApplicationDto ToApplicationDto(RecruitmentApplication application)
-    {
-        var status = NormalizeApplicationStatus(application.ApplicationStatus) ?? ApplicationPending;
-
-        return new RecruitmentApplicationDto
-        {
-            Id = application.ApplicationId,
-            RecruitId = application.RecruitId,
-            RecruitTitle = application.Recruitment?.Title ?? $"招募 {application.RecruitId}",
-            ClubId = application.Recruitment?.ClubId ?? 0,
-            ClubName = application.Recruitment?.Club?.ClubName ?? "未知社团",
-            UserId = application.UserId,
-            ApplicantName = DisplayUser(application.User) ?? $"用户 {application.UserId}",
-            StudentNo = application.User?.StudentNo,
-            ApplicationReason = application.ApplicationReason ?? string.Empty,
-            InterviewScore = application.InterviewScore is null ? null : Convert.ToDouble(application.InterviewScore.Value),
-            ApplicationStatus = ToApplicationStatusEnum(status),
-            ApplicationStatusText = ApplicationStatusText(status),
-            ReviewerUserId = application.ReviewerUserId,
-            ReviewerName = DisplayUser(application.Reviewer),
-            SubmittedAt = application.SubmittedAt,
-            ReviewedAt = application.ReviewedAt
-        };
-    }
-
-    private static string? ValidateCreateRecruitmentRequest(CreateRecruitmentRequest req)
-    {
-        if (req.CurrentUserId <= 0) return "请选择当前操作用户。";
-        if (req.ClubId <= 0) return "请选择发布招募的社团。";
-        return ValidateRecruitmentState(req.Title, req.StartAt, req.EndAt, req.Quota, req.Requirements);
-    }
-
-    private static string? ValidateRecruitmentState(
-        string? title,
-        DateTime? startAt,
-        DateTime? endAt,
-        int? quota,
-        string? requirements)
-    {
-        if (string.IsNullOrWhiteSpace(title)) return "招募标题不能为空。";
-        if (startAt is null || startAt.Value == default) return "招募开始时间不能为空。";
-        if (endAt is null || endAt.Value == default) return "招募结束时间不能为空。";
-        if (endAt.Value <= startAt.Value) return "招募结束时间必须晚于开始时间。";
-        if (quota is null or <= 0) return "招募人数必须大于 0。";
-        if (string.IsNullOrWhiteSpace(requirements)) return "招募要求不能为空。";
-        return null;
-    }
-
-    private static bool CanViewRecruitment(User viewer, Recruitment recruitment)
-    {
-        var status = NormalizeRecruitmentStorageStatus(recruitment.RecruitStatus) ?? RecruitmentStatuses.Draft;
-        if (status == RecruitmentStatuses.Draft) return CanManageRecruitment(viewer, recruitment.ClubId);
-        if (status == RecruitmentStatuses.PendingReview)
-        {
-            return CanManageRecruitment(viewer, recruitment.ClubId) || CanReviewRecruitment(viewer, recruitment);
-        }
-
-        return true;
-    }
-
-    private static bool CanManageRecruitment(User user, int clubId) =>
-        UsersController.IsSystemAdmin(user) || HasClubRecruitmentManagerRole(user, clubId);
-
-    private static bool HasClubRecruitmentManagerRole(User user, int clubId) =>
-        user.UserRoles.Any(ur =>
-            ur.ClubId == clubId &&
-            IsRecruitmentManagerRole(ur.Role));
-
-    private static bool IsOwnRecruitmentProposal(User user, Recruitment recruitment) =>
-        HasClubRecruitmentManagerRole(user, recruitment.ClubId);
-
-    private static bool CanEditRecruitment(User user, Recruitment recruitment) =>
-        NormalizeRecruitmentStorageStatus(recruitment.RecruitStatus) == RecruitmentStatuses.Draft &&
-        CanManageRecruitment(user, recruitment.ClubId);
-
-    private static bool CanDeleteDraftRecruitment(User user, Recruitment recruitment) =>
-        CanEditRecruitment(user, recruitment);
-
-    private static bool CanReviewRecruitment(User user, Recruitment recruitment) =>
-        (UsersController.IsPlatformAdmin(user) || UsersController.IsSystemAdmin(user)) &&
-        !HasClubRecruitmentManagerRole(user, recruitment.ClubId);
-
-    private static bool IsRecruitmentManagerRole(Role? role) =>
-        role is not null && RecruitmentManagerRoleCodes.Contains(Normalize(role.RoleCode));
-
-    private static bool IsClubMembershipRole(Role? role) =>
-        role is not null && ClubMembershipRoleCodes.Contains(Normalize(role.RoleCode));
-
-    private static bool CurrentUserIsMemberOfClub(User viewer, int clubId) =>
-        viewer.ClubMemberships.Any(m => m.ClubId == clubId && IsCurrentMemberRecord(m)) ||
-        viewer.UserRoles.Any(ur => ur.ClubId == clubId && IsClubMembershipRole(ur.Role));
-
-    private static bool IsCurrentMemberRecord(ClubMember member)
-    {
-        var today = BusinessToday();
-        return (member.MemberStatus == null || member.MemberStatus == MemberActive) &&
-               (member.TermStart == null || member.TermStart.Value.Date <= today) &&
-               (member.TermEnd == null || member.TermEnd.Value.Date >= today);
-    }
-
-    private static bool IsMaintainableClub(Club club) =>
-        club.AuditStatus == ClubApproved && club.ClubStatus == ClubActive;
-
-    private static string EffectiveRecruitmentStatus(Recruitment recruitment, DateTime now)
-    {
-        var storageStatus = NormalizeRecruitmentStorageStatus(recruitment.RecruitStatus) ?? RecruitmentStatuses.Draft;
-        if (storageStatus == RecruitmentStatuses.Draft) return RecruitmentStatuses.Draft;
-        if (storageStatus == RecruitmentStatuses.PendingReview) return RecruitmentStatuses.PendingReview;
-        if (storageStatus == RecruitmentStatuses.Closed) return RecruitmentStatuses.Ended;
-
-        if (recruitment.EndAt is not null && recruitment.EndAt.Value < now) return RecruitmentStatuses.Ended;
-        if (recruitment.StartAt is not null && recruitment.StartAt.Value > now) return RecruitmentStatuses.NotStarted;
-        return RecruitmentStatuses.Accepting;
-    }
-
-    private static string? NormalizeRecruitmentStorageStatus(string? status)
-    {
-        if (string.IsNullOrWhiteSpace(status)) return null;
-
-        return Normalize(status) switch
-        {
-            RecruitmentStatuses.Draft or "草稿" => RecruitmentStatuses.Draft,
-            RecruitmentStatuses.PendingReview or "pending" or "reviewing" or "审核中" or "待审核" => RecruitmentStatuses.PendingReview,
-            RecruitmentStatuses.Published or "open" or "approved" or "报名中" or "申请中" or "发布" or "已通过" => RecruitmentStatuses.Published,
-            RecruitmentStatuses.Closed or RecruitmentStatuses.Ended or "finished" or "结束" or "已结束" => RecruitmentStatuses.Closed,
-            _ => null
-        };
-    }
-
-    private static string? NormalizeRecruitmentWorkflowStatus(CreateRecruitmentRequest.RecruitStatusEnum status) =>
-        status switch
-        {
-            CreateRecruitmentRequest.RecruitStatusEnum.DraftEnum => RecruitmentStatuses.Draft,
-            CreateRecruitmentRequest.RecruitStatusEnum.PendingReviewEnum => RecruitmentStatuses.PendingReview,
-            _ => null
-        };
-
-    private static string? NormalizeRecruitmentWorkflowStatus(string? status)
-    {
-        if (string.IsNullOrWhiteSpace(status)) return null;
-
-        return Normalize(status) switch
-        {
-            RecruitmentStatuses.Draft or "草稿" => RecruitmentStatuses.Draft,
-            RecruitmentStatuses.PendingReview or "pending" or "reviewing" or "审核中" or "待审核" => RecruitmentStatuses.PendingReview,
-            _ => null
-        };
-    }
-
-    private static string? NormalizeRecruitmentStatusFilter(string? status)
-    {
-        if (string.IsNullOrWhiteSpace(status)) return null;
-
-        return Normalize(status) switch
-        {
-            RecruitmentStatuses.Draft or "草稿" => RecruitmentStatuses.Draft,
-            RecruitmentStatuses.PendingReview or "pending" or "reviewing" or "审核中" or "待审核" => RecruitmentStatuses.PendingReview,
-            RecruitmentStatuses.NotStarted or "notstarted" or "未开始" => RecruitmentStatuses.NotStarted,
-            RecruitmentStatuses.Accepting or RecruitmentStatuses.Published or "open" or "申请中" or "报名中" => RecruitmentStatuses.Accepting,
-            RecruitmentStatuses.Ended or RecruitmentStatuses.Closed or "finished" or "结束" or "已结束" => RecruitmentStatuses.Ended,
-            _ => null
-        };
-    }
-
-    private static string? NormalizeRecruitmentReviewDecision(ReviewRecruitmentRequest.DecisionEnum decision) =>
-        decision switch
-        {
-            ReviewRecruitmentRequest.DecisionEnum.ApprovedEnum => ReviewApproved,
-            ReviewRecruitmentRequest.DecisionEnum.RejectedEnum => ReviewRejected,
-            _ => null
-        };
-
-    private static string? NormalizeRecruitmentReviewDecision(string? decision)
-    {
-        if (string.IsNullOrWhiteSpace(decision)) return null;
-
-        return Normalize(decision) switch
-        {
-            ReviewApproved or "accepted" or "approve" or "通过" or "审核通过" => ReviewApproved,
-            ReviewRejected or "reject" or "驳回" or "审核驳回" => ReviewRejected,
-            _ => null
-        };
-    }
-
-    private static string? NormalizeApplicationStatus(ReviewRecruitmentApplicationRequest.DecisionEnum decision) =>
-        decision switch
-        {
-            ReviewRecruitmentApplicationRequest.DecisionEnum.AcceptedEnum => ApplicationAccepted,
-            ReviewRecruitmentApplicationRequest.DecisionEnum.RejectedEnum => ApplicationRejected,
-            _ => null
-        };
-
-    private static string? NormalizeApplicationStatus(string? status)
-    {
-        if (string.IsNullOrWhiteSpace(status)) return null;
-
-        return Normalize(status) switch
-        {
-            ApplicationPending or "待筛选" or "待审核" => ApplicationPending,
-            ApplicationAccepted or "approved" or "录取" or "已录取" => ApplicationAccepted,
-            ApplicationRejected or "拒绝" or "未录取" => ApplicationRejected,
-            _ => null
-        };
-    }
-
-    private static string RecruitmentStatusText(string status) => status switch
-    {
-        RecruitmentStatuses.Draft => "草稿",
-        RecruitmentStatuses.PendingReview => "审核中",
-        RecruitmentStatuses.NotStarted => "未开始",
-        RecruitmentStatuses.Accepting => "申请中",
-        RecruitmentStatuses.Ended => "已结束",
-        _ => "未知"
-    };
-
-    private static RecruitmentDto.RecruitStatusEnum ToRecruitmentStatusEnum(string status) => status switch
-    {
-        RecruitmentStatuses.Draft => RecruitmentDto.RecruitStatusEnum.DraftEnum,
-        RecruitmentStatuses.PendingReview => RecruitmentDto.RecruitStatusEnum.PendingReviewEnum,
-        RecruitmentStatuses.NotStarted => RecruitmentDto.RecruitStatusEnum.NotStartedEnum,
-        RecruitmentStatuses.Accepting => RecruitmentDto.RecruitStatusEnum.AcceptingEnum,
-        RecruitmentStatuses.Ended => RecruitmentDto.RecruitStatusEnum.EndedEnum,
-        _ => RecruitmentDto.RecruitStatusEnum.DraftEnum
-    };
-
-    private static RecruitmentApplicationDto.ApplicationStatusEnum ToApplicationStatusEnum(string status) => status switch
-    {
-        ApplicationPending => RecruitmentApplicationDto.ApplicationStatusEnum.PendingEnum,
-        ApplicationAccepted => RecruitmentApplicationDto.ApplicationStatusEnum.AcceptedEnum,
-        ApplicationRejected => RecruitmentApplicationDto.ApplicationStatusEnum.RejectedEnum,
-        _ => RecruitmentApplicationDto.ApplicationStatusEnum.PendingEnum
-    };
-
-    private static string ApplicationStatusText(string status) => status switch
-    {
-        ApplicationPending => "待筛选",
-        ApplicationAccepted => "已录取",
-        ApplicationRejected => "未录取",
-        _ => "未知"
-    };
-
-    private static DateTime BusinessNow() =>
-        TimeZoneInfo.ConvertTimeFromUtc(DateTime.UtcNow, BusinessTimeZone);
-
-    private static DateTime BusinessToday() =>
-        TimeZoneInfo.ConvertTimeFromUtc(DateTime.UtcNow, BusinessTimeZone).Date;
-
-    private static DateTime BusinessDate(DateTime utcDateTime)
-    {
-        var utc = utcDateTime.Kind == DateTimeKind.Utc
-            ? utcDateTime
-            : DateTime.SpecifyKind(utcDateTime, DateTimeKind.Utc);
-        return TimeZoneInfo.ConvertTimeFromUtc(utc, BusinessTimeZone).Date;
-    }
-
-    private static TimeZoneInfo ResolveBusinessTimeZone()
-    {
-        try
-        {
-            return TimeZoneInfo.FindSystemTimeZoneById("China Standard Time");
-        }
-        catch (TimeZoneNotFoundException)
-        {
-            return TimeZoneInfo.FindSystemTimeZoneById("Asia/Shanghai");
-        }
-        catch (InvalidTimeZoneException)
-        {
-            return TimeZoneInfo.FindSystemTimeZoneById("Asia/Shanghai");
-        }
-    }
-
-    private static string? DisplayUser(User? user)
-    {
-        if (user is null) return null;
-        if (!string.IsNullOrWhiteSpace(user.RealName)) return user.RealName;
-        if (!string.IsNullOrWhiteSpace(user.Username)) return user.Username;
-        return $"用户 {user.UserId}";
-    }
-
-    private static string Normalize(string? value) =>
-        (value ?? string.Empty).Trim().ToLowerInvariant();
-
-    private static string? EmptyToNull(string? value) =>
-        string.IsNullOrWhiteSpace(value) ? null : value.Trim();
 }
